mindustry, mindustry-server: Tweak manifests to follow others
Both manifests now use a launcher script to start the game,
as Start Menu entries cannot always be used to run a .jar file
directly. This causes a command prompt to appear when starting
the client, but it can be safely closed after starting the game
thanks to its use of `javaw` (instead of `java`).

To be consistent with the Nuncabola manifest, only OpenJDK is now
suggested (it should work just fine, no need to go for Oracle JDK).