Create sabaki.json
Sabaki serves as a go board and is a popular tool among go players.
We are not entirely sure if the autoupdate section works and would like to ask a sanity check on it. When  we try to run `checkver.ps1 sabaki <dir> -u` on an older version we get `ERROR You cannot call a method on a null-valued expression.` Don't know what to make of it.