flycast: Update to version 2.1
Version 2.1 uses a capitalized V in the version tag, unlike previous releases.
This caused autoupdate to fail.

I changed the autoupdate url to use a capital V, ran checkver, then reverted it
to an uncapitalized v, since I assume this was a one-time mistake, and future
releases will use the uncapitalized letter again.