Fix ScummVM installation path and ini file
ScummVM has changed the structure of the zip, the app is now in
a folder inside the zip. Add a pre_install step to move the contents
of this folder into the installation directory.

Also fix the persistence of the ini file, it was being created as
a folder and not a file.

Fixes #1073 #1094