OpenTTD: fix architecture and remove baseset installation
Baseset installation could be removed because the basesets could be downloaded afterwards, and the game stores them in the user documents folder, in which it is the default content storage folder. Plus, the user (e.g. many who have the original TTD graphics, musics and sounds OR they installed the game before) may already have the basesets ready so they don't need extra basesets. Letting the user to download OpenGFX, OpenMSX and OpenSFX for an extra time does not make sense to me.