duckstation, duckstation-preview: Fix shortcut issue, update BIOS wiki link (#1283)
Duckstation doesn't want to open in a junction folder, so I made its shortcut point to the original folder.

I also used the archive.org link for the BIOS wiki page. I would have used a different wiki or comparable link,
but could not find anything with a quick search.